CardsHub/requirements.txt
pandas
streamlit
streamlit-lottie
plotly
python-dotenv
groq
langchain_groq